Finished Login page and functionality
Changed the filetype of the login page from ejs to html since embedded was not needed after all. As well as added a javascript file for the login page that  handles the logic of logging in after retrieving the information from the server from which I added a new api call to call the added function in the database file to retrieve the selected account info to confire the given passcode and login.