Files edited: Search.css, Search.html, Search.js, ServerMain.js, DataBase.js, ItemPage.css, ItemPage.html, ItemPage.js, Business.css, Backend_NodeJS_Files/Search.js
Got rid of the ItemPage files by just integrating all of their functionality into the search page, then I added the required functions into the DataBase.js file to finish functionality of the individual item lookup when it is clicked. Edited ServerMain.js and got rid of the old ItemPage code. As well as some database modifications to accomodate the new information needed for the individual item lookup.