Small tweaks
Added back buttons to the signup and login pages as well as changed the background of the signup page to match the rest of the pages aside from the mainpage. Also fixed the code for the sidebar in the search page so it no longer looks a bit sketchy when opening and closing.