Files changed:Backend_NodeJS_Files/Search.js, Search.css, Search.js ServerMain.js
Finished the part of the search page that appears when an item is clicked, it is now  in its late prototype phase. and small edits to the css file of the search page to accommodate. Finally the backend Search file was changed to use the local placeholder image instead of the one that connects to the internet, and changed the placeholder image to something a bit easier to see on the current background.